chore: move clang-tidy config file to project directory
clang-tidy when the config is not explicitly mentioned in the command looks for it in the project directory with the name `.clang-tidy`.
And editors like VS Code recognizes the `.clang-tidy` file when code analysis is enabled.
This is not a fix for a bug. This is just trying to make clang-tidy config and usage more visible.

Signed-off-by: Balakrishna Avulapati <bavulapati@gmail.com>